pyproject.toml
[project]
name = "crew_gherkin"
version = "0.1.0"
description = "crew_gherkin using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.121.0,<1.0.0",
    "python-dotenv>=1.0.1",
    "langchain-openai>=0.1.7",
    "langchain-groq>=0.1.3"
]

[project.scripts]
crew_gherkin = "crew_gherkin.main:run"
run_crew = "crew_gherkin.main:run"
train = "crew_gherkin.main:train"
replay = "crew_gherkin.main:replay"
test = "crew_gherkin.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
